requirements.txt
python-dotenv==0.15.0
confluent-kafka==1.6.1
boto3==1.17.69